fix things up

+++ src/controllers/job.js
@@ -2,15 +2,17 @@ const jobModels = require('../models/job')
 const url = require('url')
 const uuid4 = require('uuid/v4')
 const redis = require('../helpers/redis')
+const validator = require('../helpers/validation')
+const queryString = require('querystring')
 
 module.exports = {
-  getJobs: (req, res) => {
-    const { page } = req.params
-
+  getJobs: async (req, res) => {
     /* get and set default value for orderby and order */
-    let { orderby, order } = req.query
-    if (orderby === undefined) orderby = 'date_updated'
-    if (order === undefined) order = 'ASC'
+    let { orderby, order, page } = req.query
+
+    orderby = validator.setDefaultValue(orderby, false, 'name' || 'date_updated' || 'category', 'date_updated')
+    order = validator.setDefaultValue(order, false, 'ASC' || 'DESC', 'ASC')
+    page = await validator.setDefaultValue(page, true, undefined, '1')
 
     /* get and set default value for search query(qname, qcompany) */
     let { qname, qcompany } = req.query
@@ -21,25 +23,99 @@ module.exports = {
     else qcompany = `%${qcompany}%`
     // create redis key depends on the parameter
     const redisKey = req.url
+    const host = req.hostname
 
     redis.client.get(redisKey, (error, result) => {
       if (error) {
         console.log(error)
-        throw error
-      } else if (result) {
-        result = JSON.parse(result)
-        res.json({
-          page,
-          result
+      }
+      if (result) {
+        redis.client.get('/jobs/total_data', (err, total) => {
+          if (err) {
+            console.log(err)
+            throw err
+          } else if (total) {
+            let totalPages = parseInt(total / 3)
+            if (total % 3 > 0) { totalPages += 1 }
+            result = JSON.parse(result)
+
+            let nextPage = ''
+            if (totalPages <= page) {
+              nextPage = ''
+            } else {
+              let nextQueryData = { page: 1 }
+              if (req.query.page) { nextQueryData = Object.assign({}, req.query) }
+              nextQueryData.page = parseInt(nextQueryData.page) + 1
+              nextPage = `${req.protocol}://${host}:${process.env.PORT}/job/?${queryString.stringify(nextQueryData)}`
+            }
+
+            let prevPage = ''
+            if (page <= 1) {
+              prevPage = ''
+            } else {
+              let prevQueryData = { page: 0 }
+              if (req.query.page) { prevQueryData = Object.assign({}, req.query) }
+              prevQueryData.page = parseInt(prevQueryData.page) - 1
+              prevPage = `${req.protocol}://${host}:${process.env.PORT}/job/?${queryString.stringify(prevQueryData)}`
+            }
+
+            res.json({
+              success: true,
+              message: 'Data found',
+              page,
+              result,
+              total_result: total,
+              total_pages: totalPages,
+              prev_page: prevPage,
+              next_page: nextPage
+            })
+          }
         })
       } else {
         jobModels.getJobs(page, orderby, order, qname, qcompany)
           .then((result) => {
-            redis.client.setex(redisKey, 3600, JSON.stringify(result))
-            res.json({
-              page,
-              result
-            })
+            jobModels.getTotalCountJobs(qname, qcompany)
+              .then((totalJobs) => {
+                redis.client.setex(redisKey, 3600, JSON.stringify(result))
+                redis.client.setex('/jobs/total_data', 3600, String(totalJobs[0].total_data))
+
+                let totalPages = parseInt(totalJobs[0].total_data / 3)
+                if (totalJobs[0].total_data % 3 > 0) { totalPages += 1 }
+
+                let nextPage = ''
+                if (totalPages <= page) {
+                  nextPage = ''
+                } else {
+                  let nextQueryData = { page: 1 }
+                  if (req.query.page) { nextQueryData = Object.assign({}, req.query) }
+                  nextQueryData.page = parseInt(nextQueryData.page) + 1
+                  nextPage = `${req.protocol}://${host}:${process.env.PORT}/job/?${queryString.stringify(nextQueryData)}`
+                }
+
+                let prevPage = ''
+                if (page <= 1) {
+                  prevPage = ''
+                } else {
+                  let prevQueryData = { page: 0 }
+                  if (req.query.page) { prevQueryData = Object.assign({}, req.query) }
+                  prevQueryData.page = parseInt(prevQueryData.page) - 1
+                  prevPage = `${req.protocol}://${host}:${process.env.PORT}/job/?${queryString.stringify(prevQueryData)}`
+                }
+
+                res.json({
+                  success: true,
+                  message: 'Data found',
+                  page,
+                  result,
+                  total_result: totalJobs[0].total_data,
+                  total_pages: totalPages,
+                  prev_page: prevPage,
+                  next_page: nextPage
+                })
+              })
+              .catch((err) => {
+                console.log(err)
+              })
           })
           .catch((err) => {
             console.log(err)
@@ -53,7 +129,7 @@ module.exports = {
   */
   redirectFirstPage: (req, res) => {
     const data = req.query
-    res.redirect(url.format({ pathname: '../../job/jobs/1', query: data }))
+    res.redirect(url.format({ pathname: '../../job/jobs/', query: data }))
   },
 
   // get single job
@@ -81,7 +157,9 @@ module.exports = {
     redis.getKeys('/jobs/*')
       .then(result => {
         keys = result
-        redis.client.del(keys)
+        if (keys) {
+          redis.client.del(keys)
+        }
       })
       .catch(err => {
         console.log(err)
@@ -110,7 +188,9 @@ module.exports = {
     redis.getKeys('/jobs/*')
       .then(result => {
         keys = result
-        redis.client.del(keys)
+        if (keys) {
+          redis.client.del(keys)
+        }
       })
       .catch(err => {
         console.log(err)
@@ -128,7 +208,7 @@ module.exports = {
       })
   },
 
-  deleteJob: async (req, res) => {
+  deleteJob: (req, res) => {
     const { id } = req.params
 
     /* Delete all job cache in redis because of job change */
@@ -137,7 +217,9 @@ module.exports = {
     redis.getKeys('/jobs/*')
       .then(result => {
         keys = result
-        redis.client.del(keys)
+        if (keys) {
+          redis.client.del(keys)
+        }
       })
       .catch(err => {
         console.log(err)
@@ -147,7 +229,7 @@ module.exports = {
       .then((result) => {
         res.json({
           message: 'success',
-          result
+          id
         })
       })
       .catch((err) => {
